[Feature #2178134] Refactor code to eliminate dubication
