+++ .gitignore
@@ -13,6 +13,7 @@ pids
 
 # Directory for instrumented libs generated by jscoverage/JSCover
 lib-cov
+.nyc_output
 
 # Coverage directory used by tools like istanbul
 coverage
@@ -36,6 +37,7 @@ dist/
 # Dependency directories
 node_modules/
 jspm_packages/
+.vscode
 
 # TypeScript v1 declaration files
 typings/
