--- server/Models/passengers.js
@@ -1,19 +1,0 @@
-
-class Passengers {
-  constructor(username, password, email) {
-    this.username = username;
-    this.password = password;
-    this.email = email;
-  }
-  getUsername() {
-    return this.username;
-  }
-  getEmail() {
-    return this.emial;
-  }
-  getPassword() {
-    return this.password;
-  }
-}
-
-export default Passengers;
